==> Scripts/2 Importing libraries and Python data types.ipynb <==
# 01.Importing libraries

#Import libraries
import pandas as pd
import numpy as np
import os

# 02.Data Types

#integer
x = 33
y = 67

x + y

25 + 75

x = x + y

x

#Float
a = 1.234
b = 2.345
c = a / b
print(c)

#String
d = 'Daniel '
e = 'Nuñez '
f = 'Ortiz, '
g =  'future Data Analyst'
h = d + e + f + g
print(h)

a = 'today '
b = 'is '
c = 'a '
d = 'good '
e = 'day '
f = 'for '
g = 'BBQ.'
print(a,b,c,d,e,f,g)
a = 'Beef, '
b = 'pork '
c = 'or '
d = 'chicken?'
e = a + b + c + d
print(e)

#Bonus task
w = "I'm "
t = "hungry "
f = "now. "
print((w + t + f) * 100)
