==> Scripts/3 IC Data import and descriptive analysis.ipynb <==
# 01. Importing libraries

#Import libraries
import pandas as pd
import numpy as np
import os

# 02. Import data

#Import data
df = pd.read_csv(r'C:\Users\danie\Desktop\career Foundry\31-03-2023 Instacart Basket Analysis\02 Data\Original data\orders.csv', index_col = False)

# 03. Creating a folderpath variable

#Path to folder 31-03-2023 Instacart Basket Analysis
path = r'C:\Users\danie\Desktop\career Foundry\31-03-2023 Instacart Basket Analysis\02 Data\Original data'

path

#Import data with os
df = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'orders.csv'))


# 04. Exploring Data with Pandas

#Selecting top 5 in the list (default)
df.head()

#Fisrt 30 entries
df.head(30)

#last 5 entries 
df.tail()

#Dataframe shape
df.shape

df.info()

df.dtypes

df.columns

#Statistical description of the data
df.describe()

#  05. Python Data Structures: List

#Importing a portion of the Dataset
df = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'orders.csv'), nrows = 1000)

# Creating variable list of interest
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

#Import the selected columns
df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

df.head()

#Import products.csv as df_prods
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'))

#Printing the first 20 rows
df_prods.head(20)

#Printing the last 35 rows
df_prods.tail(35)

#Columns name in df_prods
df_prods.columns

#Number of rows using shape function. (rows, columns)
df_prods.shape

#df_statistical description
df_prods.describe()

"aisle_id" max value = 134

#df_prods data types
df_prods.dtypes

"department_id" data type = int64
