==> Scripts/4 Data Wrangling & Subsetting .ipynb <==
# 01. Import libraries

#Import libraries
import pandas as pd
import numpy as np
import os

#creating folderpath variable to Original data
path = r'C:\Users\danie\Desktop\career Foundry\31-03-2023 Instacart Basket Analysis\02 Data\Original data'

#Import Data sets orders.csv & products.csv
df_ords = pd.read_csv(os.path.join(path, 'orders.csv'))
df_prods = pd.read_csv(os.path.join(path, 'products.csv'))

# 0.2 Drop function, 'df'.drop()

#Dropping 'eval_set' column
df_ords.drop(columns = ['eval_set'])

# 03. Overwriting the data frame

#Delete 'eval_set' and overwiting the data frame
df_ords = df_ords.drop(columns = ['eval_set'])

df_ords['days_since_prior_order'].value_counts(dropna = False)

#Rename a column
df_ords.rename(columns = {'order_dow' : 'orders_day_of_the_week'}, inplace = True)

df_ords.head()

#Data type change
df_ords['order_id'] = df_ords['order_id'].astype('str')

df_ords['order_id'].dtype

#Importing data set departments.csv
df_dep = pd.read_csv(os.path.join(path,'departments.csv'), index_col = False)

df_dep.head()

#Transposing df_dep
df_dep.T

#Creating a new data frame: df_dep_t
df_dep_t = df_dep.T

df_dep_t

#Adding an index to df_dep_t
df_dep_t.reset_index()

#Creating a new header
new_header = df_dep_t.iloc[0]

new_header

#Creating new df copying rows beyond the first row
df_dep_t_new = df_dep_t[1:]

df_dep_t_new

#Adding the new header
df_dep_t_new.columns = new_header

df_dep_t_new

# Creating a dictionary

#Creating Department dictionary
data_dict = df_dep_t_new.to_dict('index')

data_dict

df_prods.head()

print(data_dict.get('19'))

# Subsetting

#Creating a subset df_snack from df_prods
df_snacks = df_prods[df_prods['department_id']==19]

df_snacks

#That's the values in 'department_id' equal 19? (index, boolean)
df_prods['department_id']==19

#Selecting the rows u observations with 'department_id' equal 19
df_prods[df_prods['department_id']==19]

#Creating a new data frame with the filter above as df_snacks
df_snacks = df_prods[df_prods['department_id']==19]


df_snacks.head()

#Creating the df obeve using .loc
df_snacks_2 = df_prods.loc[df_prods['department_id'] == 19]

df_snacks_2

#creating a df with .isin([])
df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19])]

df_snacks_3

path_prep_data = path = r'C:\Users\danie\Desktop\career Foundry\31-03-2023 Instacart Basket Analysis\02 Data\Prepared data'

# Wrangling procedures part 2

df_ords.columns

df_ords

'orders_day_of_the_week' values doesn't need to be included in the analysis as a numeric variable.

#Replacing values in 'orders_day_of_the_week' for the name of the day of the week, creating 'df_ords_recoded'
df_ords_recoded = df_ords.replace(to_replace={'orders_day_of_the_week': {0:'Saturday', 1:'Sunday', 2:'Monday', 3:'Tuesday', 4:'Wednesday', 5: 'Thursday', 6:'Friday'}})

df_ords_recoded

#Replacing NaN values in column 'days_since_prior_order' with '0' using ".replace" and casting 'NaN' with "np.nan" turn
df_ords_recoded = df_ords_recoded.replace(to_replace={'days_since_prior_order': {np.nan: 0}})

df_ords_recoded

#Renaming column "days_since_prior_order" to "days_since_previous_order" without overwriting the original dataframe
df_ords_recoded.rename(columns = {'days_since_prior_order' : 'days_since_previous_order'}, inplace = False)

#Counting the amount of orders in column "order_hour_of_day using ".value_count" function
df_ords_recoded['order_hour_of_day'].value_counts(dropna = False)

The busiest time of the day is at 10 AM with 288418 orders \ 
The quiest time of the day is at 3 AM with 5474 orders

#Casting the meaning of "4" in column "department_id" from the data dictionary created previously
data_dict.get('4')

data_dict

#Creating the subset 'df_breakfast' with breakfast department data only
df_breakfast = df_prods[df_prods['department_id']==14]

df_breakfast

data_dict

#Subset 'df_dinner_parties' with the product asked for the task. Using .loc[] function and '.isin' function to cast the necessary department list.
df_dinner_parties = df_prods.loc[df_prods['department_id'].isin([5,7,12,20])]

df_dinner_parties

#Finding the amount of observations using '.shape'. Results('rows'(observations), 'columns')
df_dinner_parties.shape

#Extracting information about 'user_id = 1' from df_ords
df_ords[df_ords['user_id']==1]

#Subset df_user_1 from df_ords
df_user_1 = df_ords[df_ords['user_id']==1]

#Statistical summary
df_user_1.describe()

With data set with purchase characteristics connected with order_id or user_id, I could obtain more exhaustive information about the users' behaviors.

df_ords_recoded.to_csv(os.path.join(path_prep_data, 'orders_wrangled.csv'))

df_dep_t_new.to_csv(os.path.join(path_prep_data, 'departments_wrangled.csv'))
