==> Scripts/4 Task student example.ipynb <==
# importing libraries
import pandas as pd
import numpy as np
import os

# Data Wrangling Procedures

# importing orders data
df_ords = pd.read_csv(r'C:\Users\Insel\rchu\Instacart Basket Analysis\02 Data\Original Data\orders.csv', index_col = False)

df_ords.head()

# dropping eval_set from orders.csv
df_ords = df_ords.drop(columns = ['eval_set'])

# check if the eval_set is dropped
df_ords.head()

# displaying missing value in days_since_prior_order column 
df_ords['days_since_prior_order'].value_counts(dropna = False)

# renaming order_dow column 
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

# checking the new column
df_ords.head()

# changing data type of order_id column
df_ords['order_id'] = df_ords['order_id'].astype('str')

# checking data type of oder_id column
df_ords['order_id'].dtype

path = r'C:\Users\Insel\rchu\Instacart Basket Analysis'

# importing department.csv
df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

# checking columns of departments.csv
df_dep.head()

# transforming df_dep
df_dep.T

# creating a new dataframe
df_dep_t = df_dep.T

df_dep_t

# adding an index 
df_dep_t.reset_index()

# take the first row of df_dpt_t for the header
new_header = df_dep_t.iloc[0]

new_header

# take the data under the header row for a new df
df_dep_t_new = df_dep_t[1:]

df_dep_t_new

# set new_header for new df
df_dep_t_new.columns = new_header

df_dep_t_new

# create a data dictionary
data_dict = df_dep_t_new.to_dict('index')

data_dict

# importing products dtata (using os library)
df_prods =  pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

df_prods.head()

print(data_dict.get('19'))

# check values within df_prods that have a "department_id" of 19 and create a new df called df_snacks
df_snacks = df_prods[df_prods['department_id']==19]

df_snacks.head()

# Task step 2-13

# 2. changing data type of user_id column
df_ords['user_id'] = df_ords['user_id'].astype('str')

# checking data type of user_id column
df_ords['user_id'].dtype

# 3. change column name without overwriting the data frame.
df_ords.rename(columns = {'days_since_prior_order' : 'days_since_previous_order'}, inplace = False)

# 4. The busiest hour is for placing orders
df_ords['order_hour_of_day'].value_counts(dropna = False)

The busiest hour for palcing orders is 10 am (288418)

# 5. Determine the meaning behind a value of 4 in the "department_id" column within the df_prods dataframe using a data dictionary.
print(data_dict.get('4'))

# 6. Create a subset about breakfast item sales.
df_breakfast = df_prods.loc[df_prods['department_id'] == 14]

df_breakfast

# 7. items from the following departments: alcohol, deli, beverages, and meat/seafood
df_dinner = df_prods.loc[df_prods['department_id'].isin([5,7,12,20])]

df_dinner

8. How many rows does the last dataframe you created have?
7650 rows × 5 columns

# 9. Extract all the information about this user with a "user_id" of “1.
df_ords_user1 = df_ords[df_ords['user_id']=='1']

df_ords_user1

# 10. basic stats
df_ords_user1.describe()

# 12. Export df_ords
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

# 13. Export df_dep_t_new
df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.csv'))
