==> Scripts/5 Data Consistency Checks.ipynb <==
# 4.5: Data Consistency Checks

#Import libraries
import pandas as pd
import numpy as np
import os

# Importing Data 
df_prods = pd.read_csv(r'C:\Users\danie\Desktop\career Foundry\31-03-2023 Instacart Basket Analysis\02 Data\Original data\products.csv')

df_ords = pd.read_csv(r'C:\Users\danie\Desktop\career Foundry\31-03-2023 Instacart Basket Analysis\02 Data\Prepared data\orders_wrangled.csv')

df_ords.describe()

# Create a dataframe
df_test = pd.DataFrame()

# Create a mixed type column
df_test['mix'] = ['a', 'b', 1, True]

df_test.head()

# Checking for mixed values
for col in df_test.columns.tolist():
    weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (df_test[weird]) > 0:
        print(col)

#Turning mixed values in to strings
df_test['mix'] = df_test['mix'].astype('str')

# Finding missing values
df_prods.isnull().sum()

df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

#Using '.shape'
df_prods.shape

# Creatind df_prods_clean filtering NULL values
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean.shape

#Looking for full duplicates in a dataframe
df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

# Dropping duplicates in df_prods
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

#Path to Prepared data folder
path_prep_data = r'C:\Users\danie\Desktop\career Foundry\31-03-2023 Instacart Basket Analysis\02 Data\Prepared data'

#Exporting changes into Prepared data folder as 'products_checked.csv'
df_prods_clean_no_dups.to_csv(os.path.join(path_prep_data,'products_checked.csv'))

# Data consistency checks part 2

# 2. Running df.decribe() on df_ords
df_ords.describe()

An 'Unnamed' column appears to have the same amount of observations as the rest of the data set, and the MIN, MAX, and Quartiles do not match the 'order_id' statistical description. Therefore, further search is necessary to identify the meaning of the values in the "Unnamed" column and verify if the values in the given column are correct(or are not correct).
'order_id' and 'order_number' could lead to confusion. More self-descriptive column names or further information would be necessary for the analysis

# 3. Check for mixed types in df_ords
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

col

Mixed types found in column 'days_since_prior_order'

# 4. Fix 'Mixed types' in column 'days_since_prior_order' 
df_ords['days_since_prior_order'] = df_ords['days_since_prior_order'].astype('int64')

# 4.b Check for mixed types in df_ords
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

No mixed types found.

# 5. Check for missing values in df_ords
df_ords.isnull().sum()

No missing values was found. It's a very straigth-forward method to use '.isnull' function search for null values, and by adding 'sum()' would display the amount of null values in each column.

# 6. Addressed missing values.

6. I have found missing values (NaN) in the column 'days_since_prior_order' in Task 4.4, and I have addressed the missing values by imputing the value '0' instead using '.replace().' NaN values appear to be the first order by each 'user_id,' so replacing it with the integer '0' helps to increase consistency and prevent errors in further statistical analyses.

# 7. Check duplicates in df_ords
df_ords_dups = df_ords[df_ords.duplicated()]

df_ords_dups

No duplicated entries were found. The '.duplicated()' function scans the entire data frame searching for full-duplicated entries.

df_ords.to_csv(os.path.join(path_prep_data,'orders_checked.csv'))
