==> Scripts/6 Combining & Exporting Data (Part 1).ipynb <==
# 4.6: Combining & Exporting Data

### 01. Import libraries (pandas, Numpy, Os)

#Import libraries
import pandas as pd
import numpy as np
import os

### 02. Create paths to import & export the data.

# Create path to Original data folder
path_original = r'C:\Users\danie\Desktop\career Foundry\31-03-2023 Instacart Basket Analysis\02 Data\Original data'

# Create path to Prepared data folder
path_prepared = r'C:\Users\danie\Desktop\career Foundry\31-03-2023 Instacart Basket Analysis\02 Data\Prepared data'

### 03. Import DataFrames 

# Importing new data as 'df_ords_prior. Import data set 'df_ords', 'df_prods',
df_ords_prior = pd.read_csv(os.path.join(path_original, 'order_products__prior.csv'))
df_ords = pd.read_csv(os.path.join(path_prepared, 'orders_wrangled.csv'))
df_prods = pd.read_csv(os.path.join(path_prepared, 'products_checked.csv'))

### 04. Check DataFrame dimentions

# Check df_ords_prior '.head()'
df_ords_prior.head()

# Check df_ords_prior shape
df_ords_prior.shape

# Check df_ords shape
df_ords.shape

### 05. Merge df_ords_prior & df_ords

# Merge df_ords with df_ords_prior
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

Dataframes 'df_ords' and 'df_ords_prior' are merged using the common column "order_id" and 'inner' join as default.

### 06. Check dimentions & Integrity of the new dataframe

# Check df_merged_large '.head()'
df_merged_large.head()

# Check .value_count()
df_merged_large['_merge'].value_counts()

# Check shape of the new 'df'
df_merged_large.shape

### 07. Export 'df_merged_large' dataframe in 'Pickle' format (.pkl)

# Export merged dataframe in to pickle file (.pkl)
df_merged_large.to_pickle(os.path.join(path_prepared, 'orders_products_combined.pkl'))

df_merged_large.shape
