==> Scripts/6 Combining & Exporting Data (part 2).ipynb <==
# 4.6: Combining & Exporting Data (part 2)

### 1. Import Libraries: Pandas, Numpy, Os

#Import libraries
import pandas as pd
import numpy as np
import os

### 2. Create path to 'Prepared data'

# Create path to Prepared data folder
path_prepared = r'C:\Users\danie\Desktop\career Foundry\31-03-2023 Instacart Basket Analysis\02 Data\Prepared data'

### 3. Import DataFrame, Pickle file (.pkl)

# Import .pkl file
df_merged_large = pd.read_pickle(os.path.join(path_prepared, 'orders_products_combined.pkl')) 

#### 4. Check columns & shape

# Check dataframe with '.head()'
df_merged_large.head()

# Check dataframe shape
df_merged_large.shape

The shape of the file 'orders_products_combined.pkl' imported match with the shape of the file previously exported in the last exercise.

#### 5. Import dataframe 'products' to merge & Check shape

#Import 'products_checked.csv' from 'Prepared data' folder
df_prods = pd.read_csv(os.path.join(path_prepared, 'products_checked.csv'))

df_prods.shape

#### 6. Run merge tests

#.merge test with '.., how = outer, left, right, inner join' conditions has been checked before procceding to the next step.

#### 7. Merge DataFrames

# Merge df_merged_large & df_prods
df_ords_prods_merged = df_merged_large.merge(df_prods, on = 'product_id', how = 'inner')

df_ords_prods_merged

The outcome from running a test ".merge(... how = 'outer')" show a numerous empty values that appear to be unsold items or untracked sales during the time the 'Orders' data was recorded.

Therefore, a merge with inner join is the most suitable options for the integrity of our data.

Succefully merged dataframes.
There are two unnamed columns in the df.

#### 8. Use merge flag

# Merge flag
df_ords_prods_merged['_merge'].value_counts()

##### Look for empty values

# Finding missing values
df_ords_prods_merged.isnull().sum()

No NULL values in the entire data frame.

### 9. Export the merge final DataFrame

# Export data frame in a suitable format, keep in mind the size of the data
df_ords_prods_merged.to_pickle(os.path.join(path_prepared, 'orders_products_merged.pkl'))
