==> Scripts/7 Deriving New Variables.ipynb <==
# 4.7 Deriving New Variables

### Import Libraries: Pandas, Numpy, Os

#Import libraries
import pandas as pd
import numpy as np
import os

### Create path to 'Prepared data'

# Create path to Prepared data folder
path_prepared = r'C:\Users\danie\Desktop\career Foundry\31-03-2023 Instacart Basket Analysis\02 Data\Prepared data'

### Import DataFrame, Pickle file (.pkl)

# Import .pkl
df_ords_prods = pd.read_pickle(os.path.join(path_prepared, 'orders_products_merged.pkl'))

### Check columns & shape

# Check dataframe with '.head()'
df_ords_prods.head()

# Check dataframe shape
df_ords_prods.shape

***The shape of the file 'orders_products_merged.pkl' imported match with the shape of the file previously exported in the last exercise.***

### Create subset with first 1Million rows

# Subset df with the first million rows 
df = df_ords_prods[:1000000]

df.shape

### Run merge tests

#.merge test with '.., how = outer, left, right, inner join' conditions has been checked before procceding to the next step.

### Create new variable using 'def' with conditional function 'if, elif, else'.

# using 'def' with conditional function 'if, elif, else'.
def price_label(row):
    if row['prices'] <= 5:
        return 'Low-range product'
    elif (row['prices'] > 5) and (row['prices'] <= 15):
        return 'Mid-range product'
    elif row['prices'] > 15:
        return 'High-range'
    else: 
        return 'Not enough data'

#Creating new variable 'price_range'
df['price_range'] = df.apply(price_label, axis = 1)

# .value_counts() count of entries with the same values in a column
df['price_range'].value_counts(dropna = False)

# Checking Max value in 'prices'
df['prices'].max()

### If-Statements with the loc() Function

# Using .loc() Function to add values in a new column
df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Min-range product'

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts(dropna = False)

# Defining values unsing If-Statements with the loc() Function
df_ords_prods.loc[(df_ords_prods['prices'] <= 15) & (df_ords_prods['prices'] > 5), 'price_range_loc'] = 'Min-range product'

df_ords_prods.loc[df_ords_prods['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_ords_prods.loc[df_ords_prods['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_ords_prods['price_range_loc'].value_counts(dropna = False)

df_ords_prods['orders_day_of_the_week'].value_counts(dropna = False)

***Busiest day of the week: Saturday.
Slowest day of the week: Sunday***

### If-Statements with For-Loops

# Categorize the day of the week depending of the number of orders using For_loop
import time
start_time = time.time()

result = []

for value in df_ords_prods["orders_day_of_the_week"]:
  if value == 'Saturday':
    result.append("Busiest day")
  elif value == 'Wednesday':
    result.append("Least busy")
  else:
    result.append("Regularly busy")
print("--- %s seconds ---" % (time.time() - start_time))

result

df_ords_prods['busiest_day'] = result

# Check value frecuency in "busiest_day" 
df_ords_prods['busiest_day'].value_counts(dropna = False)

# 02. Create 'busiest_days' column

# Create a new colums. If-Statements with For_Loops. “Busiest days” (plural).
# Using '.time' function for fun.
start_time = time.time()

result2 = []

for value in df_ords_prods["orders_day_of_the_week"]:
  if value == 'Saturday' or value == 'Sunday':
    result2.append("Busiest days")
  elif value == 'Wednesday' or value =='Tuesday':
    result2.append("Slowest days")
  else:
    result2.append("Regularly busy")
print("--- %s seconds ---" % (time.time() - start_time))

# Adding result2 to a new column "busiest_ days"
df_ords_prods['busiest_days'] = result2

df_ords_prods['busiest_days'].value_counts(dropna = False)

df_ords_prods.shape

***No Empty or unknown values in 'busiest_days' column.***

#Check the values of this new column for accuracy
df_ords_prods['busiest_days'].value_counts().describe()

***Observations: Total amount of orders within Slowestdays (Tuesday and Wednesday) 7624336 . 
              Total amount of orders within Busiest days (Saturday and Sunday) 1186864412. 
              Total amount of orders within Regularly busy days (Monday, Thursday and Friday) 12916111. 
              Average 10801620 orders among the three grouped days.
              Median 11864410 orders***

#Removing 'Unnamed' columns
df_ords_prods.drop(df_ords_prods.columns[df_ords_prods.columns.str.contains('unnamed',case = False)],axis = 1, inplace = True)

df_ords_prods.shape

# Identify the busiest hours of the day.
df_ords_prods['order_hour_of_day'].value_counts(dropna = False)

# Periods of time labeled “Most orders,” “Average orders,” and “Fewest orders.”
# Create a new column containing these labels called “busiest_period_of_day.”
result_per_hours = []

for value in df_ords_prods["order_hour_of_day"]:
  if value >= 9 and value <= 16:
    result_per_hours.append("Most orders")
  elif value == 23 or (value <= 6 and value >= 0):
    result_per_hours.append("Fewest orders")
  else:
    result_per_hours.append("Average orders")

# Creating a new column "busiest_period_of_day"
df_ords_prods['busiest_period_of_day'] = result_per_hours

df_ords_prods.head()

# Value Frequency in 'busiest_period_of_day'
df_ords_prods['busiest_period_of_day'].value_counts()

### Bonus task

***For a more accurate and effective method to prevent system failure, I suggest addressing the issue by identifying the busiest hours in the week, given the fact that most active hours could differ between the days of the week ***

# Sample of the busiest 42 hours in the week
df_ords_prods[['order_hour_of_day', 'orders_day_of_the_week']].value_counts(dropna = False).nlargest(42)

***While 'Thursday' is categorized as a day with "Fewer orders," in some hours appear to be among the busiest time between 12:00 and 16:00***

# Export data frame in a suitable format, keep in mind the size of the data
start_time = time.time()

df_ords_prods.to_pickle(os.path.join(path_prepared, 'df_ords_prods_4.7_.pkl'))

print("--- %s seconds ---" % (time.time() - start_time))
