==> Scripts/8 Grouping Data & Aggregating Variables Part 2.ipynb <==
# 4.8 Grouping Data & Aggregating Variables

### Import Libraries: Pandas, Numpy, Os

#Import libraries
import pandas as pd
import numpy as np
import os
import time

start_time = time.time()

### Create path to 'Prepared data'

# Create path to Prepared data folder
path_prepared = r'C:\Users\danie\Desktop\career Foundry\31-03-2023 Instacart Basket Analysis\02 Data\Prepared data'

### Import DataFrame, Pickle file (.pkl)

# Import .pkl
df_ords_prods = pd.read_pickle(os.path.join(path_prepared, '4.8.orders_products_merged.pkl'))

### Check columns & shape

# Check dataframe with '.head()'
df_ords_prods.head()

# Check dataframe shape
df_ords_prods.shape

***The shape of the file imported match with the shape of the file previously exported in the last exercise.***

## 02. 'order_number' group by 'departmet_id'

# Group 'order_number' by 'department_id' adding the 'order_number' mean for each 'department 
df_ords_prods.groupby('department_id').agg({'order_number': ['mean']})

***Q.3. We observe 21 department_ids and the max and min in the order_number's mean had change significantly. Previusly, using the first million entries only 8 out of 21 department was reflected in the data frame,therefore, we can assume that no order from the others 13 department was made or registered within the first million rows of our data set.***

## 0.4 Create 'loyalty_flag'

# Split the data in to groups base on 'user_id'
# Apply transform() function on 'order_number'
# Create 'max_order' column to place the results of the aggregation
df_ords_prods['max_orders'] = df_ords_prods.groupby(['user_id'])['order_number'].transform(np.max)

# Assign a "loyalty" label to a user ID
df_ords_prods.loc[df_ords_prods['max_orders'] > 40, 'loyalty_flag'] = 'Loyal customer'

# Assign a "loyalty" label to a user ID
df_ords_prods.loc[(df_ords_prods['max_orders'] <= 40) & (df_ords_prods['max_order'] > 10), 'loyalty_flag'] = 'Regular customers'

#Assign a "loyalty" label to a user ID
df_ords_prods.loc[df_ords_prods['max_orders'] <= 10, 'loyalty_flag'] = 'New customer'

# Check 'Loyalty_flag' frequency
df_ords_prods['loyalty_flag'].value_counts(dropna = False)

## 05. Determine if whether the prices of products purchased by loyal customers differ from those purchased by regular or new customers.

# Summary check "loyalty_flag" using groupby() & agg() functions
df_ords_prods.groupby('loyalty_flag').agg({'prices': ['mean', 'median', 'max', 'min', 'sum']})

***Before any further analysis, I observe outliers values in 'prices' variable values which could be a error imputing the data or a mayor transaction that can influence signifycantly the results therefore would not represent the general user behaviour. The identification and omision of these values for further analysis will be conducted.***

### Detecting and deleting unusuals values in 'prices' column 

# Locate the upper outliers in 'prices' column
df_ords_prods.groupby('prices').agg({'product_id': ['min']})

***There are two values, '14900' & '99999' in 'prices' that should be categorized as 'outliers'; therefore, both observations will be omitted in the analysis***

***I suspect that the Product_id 21553 with priced 14900.0 was wrongly imputted, and I assume it should be '14.9'. It must be verify before adding this entries in our studies.***

# Delete entries (rows) with values'14900' & '99999' in column 'prices' 
df_ords_prods = df_ords_prods.drop(df_ords_prods[df_ords_prods.product_id == 21553].index)

df_ords_prods = df_ords_prods.drop(df_ords_prods[df_ords_prods.product_id == 33664].index)

# Check Data set after deleting rows 
df_ords_prods.groupby('prices').agg({'product_id': ['min']})

***Both product_ids succefuly remove without compromise the integrity of the data set.---Total of 5127 observations out of 32404859 removed it represent 0.00015% of entries in the data set.***

# Statistical summary using agg() function
df_ords_prods.groupby('loyalty_flag').agg({'prices': ['mean', 'median', 'max', 'min', 'sum', 'count']})

***There is a small difference in Average spent between all three groups.
    ---Loyal customer average spend: 7.773575.
    ---New customer average spend: 7.801206.
    ---Reagular customers 7.798262.***

***All groups share the same values in Median = 7.4, Maximum = 25.0, Minimum = 1.0***

***The bigger group based in "loyalty_flag" is "Regular customers": 15874128 users. Follow by "Loyal customer": 10282763 users and "New customer": 6242841 users*** 

## 06. Create 'Spending_flag' with prices average.

# Create new column with the average spend by user_id using groupby() & agg() functions. 
# Transform(np.mean) to get the 'prices' average.
df_ords_prods['average_spend'] = df_ords_prods.groupby(['user_id'])['prices'].transform(np.mean)

df_ords_prods['average_spend'].value_counts(dropna = False).head(7)

# Creating new column 'spender_flag' to categorize the users in two groups base on 'average_spend'
df_ords_prods.loc[(df_ords_prods['average_spend'] < 10), 'spender_flag'] = 'Low spender'

df_ords_prods.loc[(df_ords_prods['average_spend'] >= 10), 'spender_flag'] = 'High spender'

# Frequency based on 'spander_flag
df_ords_prods.groupby('spender_flag').agg({'user_id': ['count']})

***Number of users categorized as 'High spender': 119719***

***Number of users categorized as 'Low spender': 32280013***

df_ords_prods.head()

## 07. Create frequency flag based on 'days_since_prior_order'

# Creating a new column 'user_median' with the median values base on 'days_since_prior_order' for a new "column flag" 
df_ords_prods['user_median'] = df_ords_prods.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

df_ords_prods['user_median'].head()

# Creating users 'frequency_flag' column 
df_ords_prods.loc[(df_ords_prods['user_median'] > 20), 'frequency_flag'] = 'Non-frequent customer'

df_ords_prods.loc[(df_ords_prods['user_median'] > 10) & (df_ords_prods['user_median'] <= 20), 'frequency_flag'] = 'Regular customer'

df_ords_prods.loc[(df_ords_prods['user_median'] < 20), 'frequency_flag'] = 'Frequent customer'

df_ords_prods.head()

# Check frequency on 'frequency_flag' column 
df_ords_prods.groupby('frequency_flag').agg({'user_id': ['count']})

print(time.time() - start_time)

#Dropping duplicated column
df_ords_prods = df_ords_prods.drop(columns = ['max_order'])

df_ords_prods.shape


#Exporting data set in .pkl format
start_time = time.time()

df_ords_prods.to_pickle(os.path.join(path_prepared, '4.8_orders_products_merged.pkl'))

print(time.time() - start_time)
