==> Scripts/9 Data with Python Part 1. Merge, Data consistency Checks.ipynb <==
### 1. Import Libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
import time

start_time = time.time()

### 2. Create paths to folders

# Create path to Prepared data folder
path_prepared = r'C:\Users\danie\Desktop\career Foundry\31-03-2023 Instacart Basket Analysis\02 Data\Prepared data'

# Create path to 'Visualizations' folder
path_visualizations = r'C:\Users\danie\Desktop\career Foundry\31-03-2023 Instacart Basket Analysis\04 Analysis\Visualizations'

# Create path to Original data folder
path_original = r'C:\Users\danie\Desktop\career Foundry\31-03-2023 Instacart Basket Analysis\02 Data\Original data'

### 3. Import Customer Data

# Import .csv file
df_customer = pd.read_csv(os.path.join(path_original, 'customers.csv'))

df_customer.head()

df_customer.shape

### 4. Wrangle the data

#### Rename the columns

#Reformating the columns using '.str.lower()'
df_customer.columns = df_customer.columns.str.lower()

#Renaming columns into self-descriptive and suitable terms
df_customer.rename(columns = {'first name' : 'first_name', 'n_dependants' : 'number_of_kids', 'surnam' : 'surname', 'fam_status' : 'family_status'}, inplace = True)

## 5. Data quality and consistency check

# Find missing values
df_customer.isnull().sum()

***NULL values found in 'first_name' column. It doesn't influence our analysis.***

#Looking for full duplicates in a dataframe
df_dups = df_customer[df_customer.duplicated()]

df_dups

***No full duplicates found in our data set***

#Searching for odd values using '.describe()
df_customer.describe()

#Check for mixed types in df_customer
for col in df_customer.columns.tolist():
  weird = (df_customer[[col]].applymap(type) != df_customer[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_customer[weird]) > 0:
    print (col)

***Mixed values found in column 'first_name'***

#Fix 'Mixed types' in column 'first_name' 
df_customer['first_name'] = df_customer['first_name'].astype('str')

#Check for mixed types in df_customer
for col in df_customer.columns.tolist():
  weird = (df_customer[[col]].applymap(type) != df_customer[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_customer[weird]) > 0:
    print (col)

### 5. Import Data frame And Merge

#Import orders_products_merged.pkl file for merge
df_ords_prods = pd.read_pickle(os.path.join(path_prepared, 'orders_products_merged.pkl'))

#Merge df_ords_prods & df_customer
df_completed_instacart = df_ords_prods.merge(df_customer, on = 'user_id', indicator = True)

### 6. Merged Data frame Integrity and Consistency check

df_completed_instacart.shape

df_completed_instacart.columns

***Index follows a logical and self_descriptive format***

#Merge check
df_completed_instacart.groupby('_merge').agg({'user_id': ['count']})

***Complete merge between both Data sets***

#Check for mixed types in df_customer
for col in df_completed_instacart.columns.tolist():
  weird = (df_completed_instacart[[col]].applymap(type) != df_completed_instacart[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_completed_instacart[weird]) > 0:
    print (col)

***No mixed data found within the same column in the new dataframe***

# Find missing values
df_completed_instacart.isnull().sum()

***'0' missing or NULL values found in df_complete_instacart Data set***

df_completed_instacart.drop(columns = ['_merge'], inplace = True)

### 7. Export the new Data Frame

#Data set 'instacart_data_set.pkl'
df_completed_instacart.to_pickle(os.path.join(path_prepared, 'instacart_data_set.pkl'))

print([time.time() - start_time, ' seconds---'])
